+++ CHANGELOG.md
@@ -1,9 +1,22 @@
 # Changelog
 
-## [UNRELEASES]
+## [0.21.0]
 
 - not storing tool messages without the usage report into DB
 - added cached token information for Gemini models
+- fixed issue with unfinished stream response text preceding a tool call
+- default agent profile for new users set to Power Tools and added Aider as the second one
+- explicit cache control with Requesty Anthropic models used only when auto caching is disabled
+- unified LLM providers settings
+- improved onboarding flow with new unified LLM providers
+- added action to open directory with logs into Settings -> Version
+- added file completion with @ mention
+- added option to open project in AiderDesk from command line
+- notification when update is ready to be installed
+- custom commands in Agent mode
+- fixed issue with maximizing window on startup
+- added description to agent profile
+- removed rule files from the context files presented to agent as they are already in system prompt
 
 ## [0.20.0]